Add UnifiedRickRoll payload
Unified Rick Roll is a prank that is intended to be run at the office or somewhere similar on multiple devices at the same time. It will run a script on the target's computer that will wait in the background until a specified time. It will then crank up the volume and rick roll the target!